+++ packages/components/package.json
@@ -97,6 +97,7 @@
     "i18next-parser": "^0.13.0",
     "invariant": "^2.2.2",
     "jest": "20.0.3",
+    "jsdom": "^11.11.0",
     "keycode": "^2.2.0",
     "node-sass": "^4.7.2",
     "npm-watch": "^0.1.6",
